+++ utils/hashsets.go
@@ -5,11 +5,11 @@ import (
 );
 
 type Hash32Set struct {
-    set map[int32]bool;
+    set map[int64]bool;
 };
 
 func NewHash32Set() *Hash32Set {
-    return &Hash32Set{make(map[int32]bool)};
+    return &Hash32Set{make(map[int64]bool)};
 };
 
 func (set *Hash32Set) AddAll(other types.HashSet) {
@@ -18,26 +18,26 @@ func (set *Hash32Set) AddAll(other types.HashSet) {
     }
 };
 
-func (set *Hash32Set) Add(i int32) bool {
+func (set *Hash32Set) Add(i int64) bool {
     _, found := set.set[i]
     set.set[i] = true
     return !found;
 };
 
-func (set *Hash32Set) Contains(i int32) (a bool) {
+func (set *Hash32Set) Contains(i int64) (a bool) {
     return a;
 };
 
-func (set *Hash32Set) GetS() map[int32]bool {
+func (set *Hash32Set) GetS() map[int64]bool {
     return set.set;
 };
 
-func (set *Hash32Set) Get(i int32) bool {
+func (set *Hash32Set) Get(i int64) bool {
     _, found := set.set[i]
     return found;
 };
 
-func (set *Hash32Set) Remove(i int32) {
+func (set *Hash32Set) Remove(i int64) {
     delete(set.set, i);
 };
